carts view save

+++ carts/views.py
@@ -1,4 +1,3 @@
-from itertools import count
 import json
 
 from django.http import JsonResponse
@@ -16,13 +15,15 @@ class CartView(View):
 
             cart, created  = Cart.objects.get_or_create(
                 user       = request.user,
-                product_id = data['product_id']
+                product_id = data['product_id'],
+                size       = data['size'],
+                price      = data['price']
             )
             if not created and cart.count < 20:
                 cart.count += 1
                 cart.save()
-
                 return JsonResponse({'message': 'SUCCESS'}, status=200)
+
             elif not created and cart.count > 19:
                 return JsonResponse({'message': 'INVALID_COUNT'}, status=200)
 
@@ -46,8 +47,8 @@ class CartView(View):
             'count'        : cart.count,
             'product_id'   : cart.product.id,
             'product_name' : cart.product.name,
-            'product_size' : cart.product.selection.size,
-            'totalPrice'   : int(cart.count * cart.product.selection.price)
+            'size'         : cart.size,
+            'totalPrice'   : int(cart.count * cart.price)
         } for cart in carts]
         return JsonResponse({'result': cart_list}, status=200)
 
@@ -57,25 +58,31 @@ class CartView(View):
             data = json.loads(request.body)
 
             if data['count'] <= 0 or data['count'] >= 21:
+
+                
                 return JsonResponse({'message': 'INVALID_COUNT'}, status=400)
 
             if Cart.objects.filter(id=cart_id).exists():
-                cart = Cart.objects.get(id=cart_id, user=request.user
+                cart = Cart.objects.get(
+                    id   =cart_id, 
+                    user =request.user
                 )
                 cart.count = data['count']
                 cart.save()
                 return JsonResponse({'message': 'COUNT_CHANGED'}, status=201)
-
             return JsonResponse({'message': 'CART_DOES_NOT_EXIT'}, status=404)
 
         except KeyError:
-            return JsonResponse({'message': 'KEY_ERROR'}, status=400)
+            return JsonResponse({'message': 'KEY_ERROR'}, status=401)
 
     @author
     def delete(self, request):
         try:
             cart_ids = request.GET.getlist('cart_ids')
             user     = request.user
+
+            if not cart_ids:
+                return JsonResponse({'message': 'LIST_EMPTY'}, status=400)
 
             Cart.objects.filter(id__in=cart_ids, user_id=user).delete()
